server/requirements.txt
flask==3.0.3
flask-cors==4.0.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
requests==2.32.3
gunicorn==22.0.0

